Add functionality to check running interviews for a candidate
A new action 'HasRunningInterviewsForCandidateAction' has been created which checks if a candidate has any running interviews. This has been implemented in place of the previous approach in the 'CreateInterviewAction'. Also, new relations and scopes have been added to the 'Candidate' and 'Interview' models to facilitate this functionality. Changes also include updates in some tests to reflect this modification.