Add JobTitle functionality in Admin section
Implemented a new functionality for managing job titles in the administration panel. This includes creating and updating job titles, along with their corresponding tests. Added requests and resources for validating and displaying the data, actions for performing the operations, data transfer objects for data structure and respective routes in the admin section.